+++ Assets/_Scripts/Gameplay Effects/Heart.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Heart : MonoBehaviour
 {
-    [SerializeField] private float health = 30;
+    [SerializeField] private int health = 30;
     [SerializeField] private float turnSpeed = 90f;
 
     // Activates when the something collides with the cheese
